+++ paperpal/communication/__init__.py
@@ -1,0 +1,1 @@
+from .communication import GmailCommunication, construct_email_body
