+++ paperpal/communication/communication.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
 import os
 import smtplib
 from datetime import datetime
