+++ paperpal/llm/__init__.py
@@ -1,0 +1,5 @@
+from .inference import (LocalCudaInference, 
+                        AnthropicInference, 
+                        OpenAIInference, 
+                        SentenceTransformerInference, 
+                        OllamaInference)
