+++ paperpal/prompt/__init__.py
@@ -1,0 +1,6 @@
+from .prompts import (research_interests_prompt, 
+                     newsletter_prompt,
+                     research_prompt,
+                     RESEARCH_INTERESTS_SYSTEM_PROMPT,
+                     NEWSLETTER_SYSTEM_PROMPT)
+from .prompting import prompt
