+++ paperpal/prompt/prompts.py
